spring-mvc-login

+++ src/main/java/com/nhnacademy/blog/topic/repository/TopicRepository.java
@@ -3,6 +3,6 @@ package com.nhnacademy.blog.topic.repository;
 import com.nhnacademy.blog.topic.domain.Topic;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TopicRepository extends JpaRepository<Topic, Long> {
+public interface TopicRepository extends JpaRepository<Topic, Integer> {
 
 }
